Cryptoswap oracles (#81)
* Add CruptoPool Swap

* Add more LP to deployment context

* Remove comments

* Parametrize Oracle deployment & Add new collaterals

* Fix test and change Curve Router place in repo

* Fix  mistake on liquidation taking maxuint and not the full debt for repaying

+++ src/tgUSD/Oracles/sDAIOracle.sol
@@ -19,7 +19,7 @@ contract sDAIOracle is IPriceOracle {
         return oracle.latestAnswer() * 10 ** (18 - oracleDecimals);
     }
 
-    function decimals() external pure returns (uint256) {
+    function decimals() external pure returns (uint8) {
         return 18;
     }
 }
